# swagger.json
{
  "components": {
    "schemas": {}
  },
  "info": {
    "title": "Docker Daemon",
    "version": "0.0.1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/": {
      "get": {
        "operationId": "IndexController.index",
        "responses": {
          "200": {
            "content": {
              "text/html; charset=utf-8": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Index",
        "tags": [
          "Index"
        ]
      }
    },
    "/api/server": {
      "get": {
        "operationId": "ServersController.list",
        "responses": {
          "200": {
            "content": {
              "text/html; charset=utf-8": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "List",
        "tags": [
          "Servers"
        ]
      },
      "post": {
        "operationId": "ServersController.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerDto"
              }
            }
          },
          "description": "CreateServerDto",
          "required": false
        },
        "responses": {
          "201": {
            "content": {
              "text/html; charset=utf-8": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Create",
        "tags": [
          "Servers"
        ]
      }
    },
    "/api/server/{idOrName}": {
      "get": {
        "operationId": "ServersController.get",
        "parameters": [
          {
            "in": "path",
            "name": "idOrName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html; charset=utf-8": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get",
        "tags": [
          "Servers"
        ]
      },
      "put": {
        "operationId": "ServersController.action",
        "parameters": [
          {
            "in": "path",
            "name": "idOrName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html; charset=utf-8": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Action",
        "tags": [
          "Servers"
        ]
      }
    },
    "/api/resources": {
      "get": {
        "operationId": "ResourcesController.info",
        "responses": {
          "200": {
            "content": {
              "text/html; charset=utf-8": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Info",
        "tags": [
          "Resources"
        ]
      }
    }
  }
}